Last updated on 07.04.2022 - 4pm

+++ server/routes/question_paper.js
@@ -4,6 +4,7 @@ const { body, validationResult } = require('express-validator')
 const router = express.Router()
 const QuestionPaper = require('../models/QuestionPapers')
 const Question = require('../models/Questions')
+const Answer = require('../models/Answers')
 const mongoose = require('mongoose')
 const ObjectId = mongoose.Types.ObjectId
 
@@ -61,12 +62,43 @@ router.post('/delete', validateTeacher, [
       return res.status(400).json({ errors: errors.array() })
     }
     // Proceed further
-    const deletedQp = await QuestionPaper.findOneAndDelete({ _id: ObjectId(req.body.qpId), uid: ObjectId(req.user.uid) })
-    const deletedQuestions = await Question.deleteMany({ uid: ObjectId(req.user.uid), qp_id: deletedQp._id })
-    if (deletedQp && deletedQuestions) {
-      return res.json({ success: 'Question paper deleted!' })
+    const requiredQP = await QuestionPaper.findOne({ _id: ObjectId(req.body.qpId), uid: ObjectId(req.user.uid) })
+    if (requiredQP) {
+      const deletedQp = await QuestionPaper.findOneAndDelete({ _id: ObjectId(req.body.qpId), uid: ObjectId(req.user.uid) })
+      const deletedQuestions = await Question.deleteMany({ uid: ObjectId(req.user.uid), qp_id: deletedQp._id })
+      const deletedAnswers = await Answer.deleteMany({ paper_id: requiredQP.paper_id })
+      if (deletedQp && deletedQuestions && deletedAnswers) {
+        return res.json({ success: 'Question paper deleted!' })
+      }
+      return res.status(400).json({ error: 'Error occurred while deleting paper.' })
     }
-    return res.status(400).json({ error: 'Error occurred while deleting paper.' })
+    return res.status(400).json({ error: 'Cannot find question paper.' })
+  } catch (error) {
+    console.error(error)
+    return res.status(500).json({ error: 'Internal Server Error' })
+  }
+})
+
+router.post('/rename', validateTeacher, [
+  body('qpId', 'Request processing error...').isLength({ min: 1 }),
+  body('name', 'Name of question paper must contain at least 5 characters').isLength({ min: 5 })
+], async (req, res) => {
+  try {
+    // Finds error in validation and return bad request
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+    // Proceed further
+    const requiredQP = await QuestionPaper.findOne({ _id: ObjectId(req.body.qpId), uid: ObjectId(req.user.uid) })
+    if (requiredQP) {
+      const renamedQP = await QuestionPaper.findOneAndUpdate({ _id: ObjectId(req.body.qpId), uid: ObjectId(req.user.uid) }, { name: req.body.name.trim() }, { new: true })
+      if (renamedQP) {
+        return res.json({ success: 'Question paper renamed successfully!', db_value: renamedQP.name })
+      }
+      return res.status(400).json({ error: 'Error occurred while renaming paper.' })
+    }
+    return res.status(400).json({ error: 'Cannot find question paper.' })
   } catch (error) {
     console.error(error)
     return res.status(500).json({ error: 'Internal Server Error' })
@@ -87,7 +119,7 @@ router.post('/get_total_no_questions', validateTeacher, [
     if (requiredQP) {
       const all_questions_ids = await Question.find({ uid: ObjectId(req.user.uid), qp_id: ObjectId(req.body.qp_id) }).select({ _id: 1, marks: 1 })
       if (all_questions_ids) {
-        return res.json({ total: all_questions_ids, paper_id: requiredQP.paper_id, duration: requiredQP.duration })
+        return res.json({ total: all_questions_ids, paper_id: requiredQP.paper_id, duration: requiredQP.duration, name: requiredQP.name })
       }
       return res.status(400).json({ error: 'Cannot find question paper.' })
     }
